Flyttter personoplysninger fra grunnlagsdata til faktagrunnlag
