--- src/main/kotlin/no/nav/tilleggsstonader/sak/opplysninger/grunnlag/GrunnlagUtil.kt
@@ -1,8 +1,0 @@
-package no.nav.tilleggsstonader.sak.opplysninger.grunnlag
-
-fun no.nav.tilleggsstonader.sak.opplysninger.pdl.dto.Navn.tilNavn() =
-    Navn(
-        fornavn = fornavn,
-        mellomnavn = mellomnavn,
-        etternavn = etternavn,
-    )
