+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårsoppsummering/VilkårsoppsummeringUtil.kt
@@ -1,7 +1,7 @@
 package no.nav.tilleggsstonader.sak.vilkår.vilkårsoppsummering
 
 import no.nav.tilleggsstonader.kontrakter.felles.Datoperiode
-import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.GrunnlagBarn
+import no.nav.tilleggsstonader.sak.opplysninger.grunnlag.faktagrunnlag.GrunnlagBarn
 import java.time.LocalDate
 
 object VilkårsoppsummeringUtil {
