+++ src/test/resources/faktaGrunnlag/PERSONOPPLYSNINGER.json
@@ -1,0 +1,25 @@
+{
+  "type": "PERSONOPPLYSNINGER",
+  "navn": {
+    "fornavn": "Fornavn",
+    "mellomnavn": "Mellomnavn",
+    "etternavn": "Etternavn"
+  },
+  "fødsel": {
+    "fødselsdato": "2025-01-01",
+    "fødselsår": 2025
+  },
+  "barn": [
+    {
+      "ident": "11111122222",
+      "navn": {
+        "fornavn": "Fornavn barn",
+        "mellomnavn": "Mellomnavn barn",
+        "etternavn": "Etternavn barn"
+      },
+      "fødselsdato": "2025-01-01",
+      "alder": 0,
+      "dødsdato": "2025-01-02"
+    }
+  ]
+}
